+++ src/xmp.cpp
@@ -953,8 +953,8 @@ namespace Exiv2 {
                 meta.SetProperty(ns.c_str(), i.tagName().c_str(), nullptr, options);
                 for (size_t idx = 0; idx < i.count(); ++idx) {
                     const std::string item = i.tagName() + "[" + toString(idx + 1) + "]";
-                    printNode(ns, item, i.toString(idx), 0);
-                    meta.SetProperty(ns.c_str(), item.c_str(), i.toString(idx).c_str());
+                    printNode(ns, item, i.toString(static_cast<long>(idx)), 0);
+                    meta.SetProperty(ns.c_str(), item.c_str(), i.toString(static_cast<long>(idx)).c_str());
                 }
                 continue;
             }
